+++ packages/astro/src/core/app/manifest.ts
@@ -129,3 +129,18 @@ export function deserializeRouteInfo(rawRouteInfo: SerializedRouteInfo): RouteIn
 		routeData: deserializeRouteData(rawRouteInfo.routeData),
 	};
 }
+
+export function queuePoolSize(
+	config: NonNullable<SSRManifest['experimentalQueuedRendering']>,
+): number {
+	return config?.poolSize ?? 1000;
+}
+export function queueContentCache(
+	config: NonNullable<SSRManifest['experimentalQueuedRendering']>,
+): boolean {
+	return config?.contentCache ?? false;
+}
+
+export function queueRenderingEnabled(config: SSRManifest['experimentalQueuedRendering']): boolean {
+	return config?.enabled ?? false;
+}