+++ packages/astro/src/vite-plugin-app/pipeline.ts
@@ -28,6 +28,9 @@ import type {
 import { getComponentMetadata } from '../vite-plugin-astro-server/metadata.js';
 import { createResolve } from '../vite-plugin-astro-server/resolve.js';
 import { PAGE_SCRIPT_ID } from '../vite-plugin-scripts/index.js';
+import { newNodePool } from '../runtime/server/render/queue/pool.js';
+import { HTMLStringCache } from '../runtime/server/html-string-cache.js';
+import { queueRenderingEnabled } from '../core/app/manifest.js';
 
 /**
  * This Pipeline is used when the Vite SSR environment is runnable.
@@ -68,6 +71,10 @@ export class RunnablePipeline extends Pipeline {
 	) {
 		const pipeline = new RunnablePipeline(loader, logger, manifest, settings, getDebugInfo);
 		pipeline.routesList = manifestData;
+		if (queueRenderingEnabled(manifest.experimentalQueuedRendering)) {
+			pipeline.nodePool = newNodePool(manifest.experimentalQueuedRendering!);
+			pipeline.htmlStringCache = new HTMLStringCache(1000); // Use default size
+		}
 		return pipeline;
 	}